fix: move constants to types file to avoid circ imports. create constant for content types with a side bar
